+++ Infinity Tower/Assets/Enemy/01_Scripts/Boss/KingSlime_PhaseChange.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+
+public partial class KingSlime
+{
+    [SerializeField]
+    private Transform NewPhasePlayerPosition;
+
+    public partial void PhaseChange(Transform PlayerPos)
+    {
+        WaitNewAction -= .5f;
+        PlayerPos.position = NewPhasePlayerPosition.position;
+        Start();
+        HP = MaxHP;
+    }
+}
