+++ Infinity Tower/Assets/Enemy/04_Prefab/Boss/KingSlime/KingSlime.prefab
@@ -204,13 +204,13 @@ MonoBehaviour:
   m_EditorClassIdentifier: Assembly-CSharp::KingSlime
   Fly: 0
   isDie: 0
-  defaultHP: 0
+  defaultHP: 200
   parentCanvas: {fileID: 0}
   HealthBar: {fileID: 1829174699335507379, guid: 459755b9939d01442bd7674b213cf12b, type: 3}
-  <hitText>k__BackingField: {fileID: 0}
-  WaitNewAction: 1.5
+  <hitText>k__BackingField: {fileID: 731421580909642978, guid: e5529f85ddade9f43b752af74f8c99dc, type: 3}
+  WaitNewAction: 3
   GroggyTime: 5
-  state: 4
+  state: 0
   AttackAreaValue:
   - {x: 25, y: 1}
   - {x: 15, y: 3}
