+++ Infinity Tower/Assets/Map/05_Prefab/HealthTextCanvas.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 531d42619e4ade44ab899cfc659fb2b0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
